different spec cache orgs work - HW works

+++ src/arch/x86/decoder.hh
@@ -487,9 +487,9 @@ protected:
 	// Interface for fetch!
     // tells fetch stage that if a speculative trace is availble for this PC
     // LVPredictor return int8_t confidence, if this confidence if less than zero then just return
-	unsigned isTraceAvailable(Addr addr, int64_t value, int8_t confidence);
+	unsigned isTraceAvailable(Addr addr, uint64_t value, uint64_t confidence);
 
-    StaticInstPtr getSuperOptimizedMicroop(unsigned traceID, X86ISA::PCState &thisPC, X86ISA::PCState &nextPC, bool &predict_taken);
+    StaticInstPtr getSuperOptimizedMicroop(uint64_t traceID, X86ISA::PCState &thisPC, X86ISA::PCState &nextPC, bool &predict_taken);
     void updateStreamTrace(unsigned traceID, X86ISA::PCState &thisPC);
 
     void regStats();
